+++ ydb/core/kqp/ut/olap/combinatory/abstract.cpp
@@ -1,0 +1,5 @@
+#include "abstract.h"
+
+namespace NKikimr::NKqp {
+
+}   // namespace NKikimr::NKqp
