+++ ydb/core/kqp/ut/olap/combinatory/execute.cpp
@@ -1,0 +1,5 @@
+#include "execute.h"
+
+namespace NKikimr::NKqp {
+
+}   // namespace NKikimr::NKqp
