+++ ydb/core/kqp/ut/olap/json_ut.cpp
@@ -1,3 +1,4 @@
+#include "combinatory/variator.h"
 #include "helpers/get_value.h"
 #include "helpers/local.h"
 #include "helpers/query_executor.h"
@@ -24,402 +25,6 @@
 namespace NKikimr::NKqp {
 
 Y_UNIT_TEST_SUITE(KqpOlapJson) {
-    class ICommand {
-    private:
-        virtual TConclusionStatus DoExecute(TKikimrRunner& kikimr) = 0;
-
-    public:
-        virtual ~ICommand() = default;
-
-        TConclusionStatus Execute(TKikimrRunner& kikimr) {
-            return DoExecute(kikimr);
-        }
-    };
-
-    class TSchemaCommand: public ICommand {
-    private:
-        const TString Command;
-        virtual TConclusionStatus DoExecute(TKikimrRunner& kikimr) override {
-            Cerr << "EXECUTE: " << Command << Endl;
-            auto session = kikimr.GetTableClient().CreateSession().GetValueSync().GetSession();
-            auto result = session.ExecuteSchemeQuery(Command).GetValueSync();
-            UNIT_ASSERT_C(result.GetStatus() == NYdb::EStatus::SUCCESS, result.GetIssues().ToString());
-            return TConclusionStatus::Success();
-        }
-
-    public:
-        TSchemaCommand(const TString& command)
-            : Command(command) {
-        }
-    };
-
-    class TDataCommand: public ICommand {
-    private:
-        const TString Command;
-        virtual TConclusionStatus DoExecute(TKikimrRunner& kikimr) override {
-            Cerr << "EXECUTE: " << Command << Endl;
-            auto session = kikimr.GetTableClient().CreateSession().GetValueSync().GetSession();
-            auto client = kikimr.GetQueryClient();
-            auto prepareResult = client.ExecuteQuery(Command, NYdb::NQuery::TTxControl::BeginTx().CommitTx()).ExtractValueSync();
-            UNIT_ASSERT_C(prepareResult.IsSuccess(), prepareResult.GetIssues().ToString());
-            return TConclusionStatus::Success();
-        }
-
-    public:
-        TDataCommand(const TString& command)
-            : Command(command) {
-        }
-    };
-
-    class TSelectCommand: public ICommand {
-    private:
-        TString Command;
-        TString Compare;
-        std::optional<ui64> ExpectIndexSkip;
-        std::optional<ui64> ExpectIndexNoData;
-        std::optional<ui64> ExpectIndexApprove;
-
-        virtual TConclusionStatus DoExecute(TKikimrRunner& kikimr) override {
-            auto controller = NYDBTest::TControllers::GetControllerAs<NYDBTest::NColumnShard::TController>();
-            AFL_VERIFY(controller);
-            const i64 indexSkipStart = controller->GetIndexesSkippingOnSelect().Val();
-            const i64 indexApproveStart = controller->GetIndexesApprovedOnSelect().Val();
-            const i64 indexNoDataStart = controller->GetIndexesSkippedNoData().Val();
-
-            const i64 headerSkipStart = controller->GetHeadersSkippingOnSelect().Val();
-            const i64 headerApproveStart = controller->GetHeadersApprovedOnSelect().Val();
-            const i64 headerNoDataStart = controller->GetHeadersSkippedNoData().Val();
-
-            Cerr << "EXECUTE: " << Command << Endl;
-            auto session = kikimr.GetTableClient().CreateSession().GetValueSync().GetSession();
-            auto it = kikimr.GetQueryClient().StreamExecuteQuery(Command, NYdb::NQuery::TTxControl::BeginTx().CommitTx()).ExtractValueSync();
-            UNIT_ASSERT_VALUES_EQUAL_C(it.GetStatus(), NYdb::EStatus::SUCCESS, it.GetIssues().ToString());
-            TString output = StreamResultToYson(it);
-            if (Compare) {
-                Cerr << "COMPARE: " << Compare << Endl;
-                Cerr << "OUTPUT: " << output << Endl;
-                CompareYson(output, Compare);
-            }
-            const ui32 iSkip = controller->GetIndexesSkippingOnSelect().Val() - indexSkipStart;
-            const ui32 iNoData = controller->GetIndexesSkippedNoData().Val() - indexNoDataStart;
-            const ui32 iApproves = controller->GetIndexesApprovedOnSelect().Val() - indexApproveStart;
-            Cerr << "INDEX:" << iNoData << "/" << iSkip << "/" << iApproves << Endl;
-
-            const ui32 hSkip = controller->GetHeadersSkippingOnSelect().Val() - headerSkipStart;
-            const ui32 hNoData = controller->GetHeadersSkippedNoData().Val() - headerNoDataStart;
-            const ui32 hApproves = controller->GetHeadersApprovedOnSelect().Val() - headerApproveStart;
-            Cerr << "HEADER:" << hNoData << "/" << hSkip << "/" << hApproves << Endl;
-            if (ExpectIndexSkip) {
-                AFL_VERIFY(iSkip + hSkip == *ExpectIndexSkip)("expect", ExpectIndexSkip)("ireal", iSkip)("hreal", hSkip)(
-                                            "current", controller->GetIndexesSkippingOnSelect().Val())("pred", indexSkipStart);
-            }
-            if (ExpectIndexNoData) {
-                AFL_VERIFY(iNoData == *ExpectIndexNoData)("expect", ExpectIndexNoData)("real", iNoData)(
-                                        "current", controller->GetIndexesSkippedNoData().Val())("pred", indexNoDataStart);
-            }
-            if (ExpectIndexApprove) {
-                AFL_VERIFY(iApproves == *ExpectIndexApprove)("expect", ExpectIndexApprove)("real", iApproves)(
-                                          "current", controller->GetIndexesApprovedOnSelect().Val())("pred", indexApproveStart);
-            }
-            return TConclusionStatus::Success();
-        }
-
-    public:
-        bool DeserializeFromString(const TString& info) {
-            auto lines = StringSplitter(info).SplitBySet("\n").ToList<TString>();
-            std::optional<ui32> state;
-            for (auto&& l : lines) {
-                l = Strip(l);
-                if (l.StartsWith("READ:")) {
-                    l = l.substr(5);
-                    state = 0;
-                } else if (l.StartsWith("EXPECTED:")) {
-                    l = l.substr(9);
-                    state = 1;
-                } else if (l.StartsWith("IDX_ND_SKIP_APPROVE:")) {
-                    state = 2;
-                    l = l.substr(20);
-                } else {
-                    AFL_VERIFY(state)("line", l);
-                }
-
-                if (*state == 0) {
-                    Command += l;
-                } else if (*state == 1) {
-                    Compare += l;
-                } else if (*state == 2) {
-                    auto idxExpectations = StringSplitter(l).SplitBySet(" ,.;").SkipEmpty().ToList<TString>();
-                    AFL_VERIFY(idxExpectations.size() == 3)("size", idxExpectations.size())("string", l);
-                    if (idxExpectations[0] != "{}") {
-                        ui32 res;
-                        AFL_VERIFY(TryFromString<ui32>(idxExpectations[0], res))("string", l);
-                        ExpectIndexNoData = res;
-                    }
-                    if (idxExpectations[1] != "{}") {
-                        ui32 res;
-                        AFL_VERIFY(TryFromString<ui32>(idxExpectations[1], res))("string", l);
-                        ExpectIndexSkip = res;
-                    }
-                    if (idxExpectations[2] != "{}") {
-                        ui32 res;
-                        AFL_VERIFY(TryFromString<ui32>(idxExpectations[2], res))("string", l);
-                        ExpectIndexApprove = res;
-                    }
-                } else {
-                    AFL_VERIFY(false)("line", l);
-                }
-            }
-            return true;
-        }
-
-        TSelectCommand() = default;
-    };
-
-    class TStopCompactionCommand: public ICommand {
-    private:
-        virtual TConclusionStatus DoExecute(TKikimrRunner& /*kikimr*/) override {
-            auto controller = NYDBTest::TControllers::GetControllerAs<NYDBTest::NColumnShard::TController>();
-            AFL_VERIFY(controller);
-            controller->DisableBackground(NKikimr::NYDBTest::ICSController::EBackground::Compaction);
-            return TConclusionStatus::Success();
-        }
-
-    public:
-        TStopCompactionCommand() {
-        }
-    };
-
-    class TOneActualizationCommand: public ICommand {
-    private:
-        virtual TConclusionStatus DoExecute(TKikimrRunner& kikimr) override {
-            {
-                auto alterQuery =
-                    TStringBuilder()
-                    << "ALTER OBJECT `/Root/ColumnTable` (TYPE TABLE) SET (ACTION=UPSERT_OPTIONS, SCHEME_NEED_ACTUALIZATION=`true`);";
-                auto session = kikimr.GetTableClient().CreateSession().GetValueSync().GetSession();
-                auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-                AFL_VERIFY(alterResult.GetStatus() == NYdb::EStatus::SUCCESS)("error", alterResult.GetIssues().ToString());
-            }
-            auto controller = NYDBTest::TControllers::GetControllerAs<NYDBTest::NColumnShard::TController>();
-            AFL_VERIFY(controller);
-            controller->WaitActualization(TDuration::Seconds(10));
-            return TConclusionStatus::Success();
-        }
-    };
-
-    class TOneCompactionCommand: public ICommand {
-    private:
-        virtual TConclusionStatus DoExecute(TKikimrRunner& /*kikimr*/) override {
-            auto controller = NYDBTest::TControllers::GetControllerAs<NYDBTest::NColumnShard::TController>();
-            AFL_VERIFY(controller);
-            AFL_VERIFY(!controller->IsBackgroundEnable(NKikimr::NYDBTest::ICSController::EBackground::Compaction));
-            const i64 compactions = controller->GetCompactionFinishedCounter().Val();
-            controller->EnableBackground(NKikimr::NYDBTest::ICSController::EBackground::Compaction);
-            const TInstant start = TInstant::Now();
-            while (TInstant::Now() - start < TDuration::Seconds(5)) {
-                if (compactions < controller->GetCompactionFinishedCounter().Val()) {
-                    Cerr << "COMPACTION_HAPPENED: " << compactions << " -> " << controller->GetCompactionFinishedCounter().Val() << Endl;
-                    break;
-                }
-                Cerr << "WAIT_COMPACTION: " << controller->GetCompactionFinishedCounter().Val() << Endl;
-                Sleep(TDuration::MilliSeconds(300));
-            }
-            AFL_VERIFY(compactions < controller->GetCompactionFinishedCounter().Val());
-            controller->DisableBackground(NKikimr::NYDBTest::ICSController::EBackground::Compaction);
-            return TConclusionStatus::Success();
-        }
-
-    public:
-        TOneCompactionCommand() {
-        }
-    };
-
-    class TWaitCompactionCommand: public ICommand {
-    private:
-        virtual TConclusionStatus DoExecute(TKikimrRunner& /*kikimr*/) override {
-            auto controller = NYDBTest::TControllers::GetControllerAs<NYDBTest::NColumnShard::TController>();
-            AFL_VERIFY(controller);
-            controller->WaitCompactions(TDuration::Seconds(5));
-            return TConclusionStatus::Success();
-        }
-
-    public:
-        TWaitCompactionCommand() {
-        }
-    };
-
-    class TBulkUpsertCommand: public ICommand {
-    private:
-        TString TableName;
-        TString ArrowBatch;
-        Ydb::StatusIds_StatusCode ExpectedCode = Ydb::StatusIds::SUCCESS;
-
-    public:
-        TBulkUpsertCommand() = default;
-
-        virtual TConclusionStatus DoExecute(TKikimrRunner& kikimr) override {
-            TLocalHelper lHelper(kikimr);
-            lHelper.SendDataViaActorSystem(TableName,
-                NArrow::TStatusValidator::GetValid(NArrow::NSerialization::TNativeSerializer().Deserialize(ArrowBatch)), ExpectedCode);
-            return TConclusionStatus::Success();
-        }
-
-        bool DeserializeFromString(const TString& info) {
-            auto lines = StringSplitter(info).SplitBySet("\n").SkipEmpty().ToList<TString>();
-            if (lines.size() < 2 || lines.size() > 3) {
-                return false;
-            }
-            TableName = Strip(lines[0]);
-            ArrowBatch = Base64Decode(Strip(lines[1]));
-            AFL_VERIFY(!!ArrowBatch);
-            if (lines.size() == 3) {
-                if (!Ydb::StatusIds_StatusCode_Parse(Strip(lines[2]), &ExpectedCode)) {
-                    return false;
-                }
-                //                if (lines[2] == "SUCCESS") {
-                //                } else if (lines[2] = "INTERNAL_ERROR") {
-                //                    ExpectedCode = Ydb::StatusIds::INTERNAL_ERROR;
-                //                } else if (lines[2] == "BAD_REQUEST") {
-                //                    ExpectedCode = Ydb::StatusIds::BAD_REQUEST;
-                //                } else {
-                //                    return false;
-                //                }
-            }
-            return true;
-        }
-    };
-
-    class TScriptExecutor {
-    private:
-        std::vector<std::shared_ptr<ICommand>> Commands;
-
-    public:
-        TScriptExecutor(const std::vector<std::shared_ptr<ICommand>>& commands)
-            : Commands(commands) {
-        }
-        void Execute() {
-            NKikimrConfig::TAppConfig appConfig;
-            appConfig.MutableTableServiceConfig()->SetEnableOlapSink(true);
-            auto settings = TKikimrSettings().SetAppConfig(appConfig).SetColumnShardAlterObjectEnabled(true).SetWithSampleTables(false);
-            TKikimrRunner kikimr(settings);
-            auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-            csController->SetOverridePeriodicWakeupActivationPeriod(TDuration::Seconds(1));
-            csController->SetOverrideLagForCompactionBeforeTierings(TDuration::Seconds(1));
-            csController->SetOverrideMemoryLimitForPortionReading(1e+10);
-            csController->SetOverrideBlobSplitSettings(NOlap::NSplitter::TSplitSettings());
-            for (auto&& i : Commands) {
-                i->Execute(kikimr);
-            }
-            AFL_VERIFY(NColumnShard::TMonitoringObjectsCounter<NOlap::NGroupedMemoryManager::TProcessMemoryScope>::GetCounter().Val() == 0)("count",
-                           NColumnShard::TMonitoringObjectsCounter<NOlap::NGroupedMemoryManager::TProcessMemoryScope>::GetCounter().Val());
-            AFL_VERIFY(NColumnShard::TMonitoringObjectsCounter<NOlap::NReader::NCommon::IDataSource>::GetCounter().Val() == 0)(
-                           "count", NColumnShard::TMonitoringObjectsCounter<NOlap::NReader::NCommon::IDataSource>::GetCounter().Val());
-        }
-    };
-
-    class TScriptVariator {
-    private:
-        std::vector<TScriptExecutor> Scripts;
-        std::shared_ptr<ICommand> BuildCommand(TString command) {
-            if (command.StartsWith("BULK_UPSERT:")) {
-                command = command.substr(12);
-                auto result = std::make_shared<TBulkUpsertCommand>();
-                AFL_VERIFY(result->DeserializeFromString(command));
-                return result;
-            } else if (command.StartsWith("SCHEMA:")) {
-                command = command.substr(7);
-                return std::make_shared<TSchemaCommand>(command);
-            } else if (command.StartsWith("DATA:")) {
-                command = command.substr(5);
-                return std::make_shared<TDataCommand>(command);
-            } else if (command.StartsWith("READ:")) {
-                auto result = std::make_shared<TSelectCommand>();
-                AFL_VERIFY(result->DeserializeFromString(command));
-                return result;
-            } else if (command.StartsWith("WAIT_COMPACTION")) {
-                return std::make_shared<TWaitCompactionCommand>();
-            } else if (command.StartsWith("STOP_COMPACTION")) {
-                return std::make_shared<TStopCompactionCommand>();
-            } else if (command.StartsWith("ONE_COMPACTION")) {
-                return std::make_shared<TOneCompactionCommand>();
-            } else if (command.StartsWith("ONE_ACTUALIZATION")) {
-                return std::make_shared<TOneActualizationCommand>();
-            } else {
-                AFL_VERIFY(false)("command", command);
-                return nullptr;
-            }
-        }
-        void BuildScripts(const std::vector<std::vector<std::shared_ptr<ICommand>>>& commands, const ui32 currentLayer,
-            std::vector<std::shared_ptr<ICommand>>& currentScript, std::vector<TScriptExecutor>& scripts) {
-            if (currentLayer == commands.size()) {
-                scripts.emplace_back(currentScript);
-                return;
-            }
-            for (auto&& i : commands[currentLayer]) {
-                currentScript.emplace_back(i);
-                BuildScripts(commands, currentLayer + 1, currentScript, scripts);
-                currentScript.pop_back();
-            }
-        }
-
-        void BuildVariantsImpl(const std::vector<std::vector<TString>>& chunks, const ui32 currentLayer, std::vector<TString>& currentCommand,
-            std::vector<TString>& results) {
-            if (currentLayer == chunks.size()) {
-                results.emplace_back(JoinSeq("", currentCommand));
-                return;
-            }
-            for (auto&& i : chunks[currentLayer]) {
-                currentCommand.emplace_back(i);
-                BuildVariantsImpl(chunks, currentLayer + 1, currentCommand, results);
-                currentCommand.pop_back();
-            }
-        }
-        std::vector<TString> BuildVariants(const TString& command) {
-            auto chunks = StringSplitter(command).SplitByString("$$").ToList<TString>();
-            std::vector<std::vector<TString>> chunksVariants;
-            for (ui32 i = 0; i < chunks.size(); ++i) {
-                if (i % 2 == 0) {
-                    chunksVariants.emplace_back(std::vector<TString>({ chunks[i] }));
-                } else {
-                    chunksVariants.emplace_back(StringSplitter(chunks[i]).SplitBySet("|").ToList<TString>());
-                }
-            }
-            std::vector<TString> result;
-            std::vector<TString> currentCommand;
-            BuildVariantsImpl(chunksVariants, 0, currentCommand, result);
-            return result;
-        }
-
-    public:
-        TScriptVariator(const TString& script) {
-            auto lines = StringSplitter(script).SplitByString("\n").ToList<TString>();
-            lines.erase(std::remove_if(lines.begin(), lines.end(),
-                            [](const TString& l) {
-                                return Strip(l).StartsWith("#");
-                            }),
-                lines.end());
-            auto commands = StringSplitter(JoinSeq("\n", lines)).SplitByString("------").ToList<TString>();
-            std::vector<std::vector<std::shared_ptr<ICommand>>> commandsDescription;
-            for (auto&& i : commands) {
-                auto& cVariants = commandsDescription.emplace_back();
-                i = Strip(i);
-                std::vector<TString> variants = BuildVariants(i);
-                for (auto&& v : variants) {
-                    cVariants.emplace_back(BuildCommand(v));
-                }
-            }
-            std::vector<TScriptExecutor> scripts;
-            std::vector<std::shared_ptr<ICommand>> scriptCommands;
-            BuildScripts(commandsDescription, 0, scriptCommands, Scripts);
-        }
-
-        void Execute() {
-            for (auto&& i : Scripts) {
-                i.Execute();
-            }
-        }
-    };
-
     Y_UNIT_TEST(EmptyVariants) {
         TString script = R"(
             STOP_COMPACTION
