+++ ydb/core/kqp/ut/olap/ya.make
@@ -13,6 +13,7 @@ ENDIF()
 SRCS(
     GLOBAL blobs_sharing_ut.cpp
     GLOBAL kqp_olap_ut.cpp
+    dictionary_ut.cpp
     aggregations_ut.cpp
     clickbench_ut.cpp
     compression_ut.cpp
@@ -39,6 +40,7 @@ PEERDIR(
     ydb/core/tx/columnshard/test_helper
     ydb/core/tx/columnshard
     ydb/core/kqp/ut/olap/helpers
+    ydb/core/kqp/ut/olap/combinatory
     ydb/core/tx/datashard/ut_common
     ydb/public/sdk/cpp/src/client/operation
 )
